Rename LinkPreviewService to HomebasePublicPageService
The service renders the full public index page for an identity (head
metadata, webfinger/DID discovery links, schema.org JSON-LD, and SSR
noscript content), not just link previews — and is about to take on
deciding whether a tenant gets a public home page at all. Rename the
class and move the Odin.Services.LinkPreview namespace/directory to
Odin.Services.PublicPage to match. No behavior change.
