+++ src/services/Odin.Services/PublicPage/HomebasePublicPageService.cs
@@ -16,20 +16,20 @@ using Odin.Core.Exceptions;
 using Odin.Core.Serialization;
 using Odin.Core.Storage.Cache;
 using Odin.Services.Base;
-using Odin.Services.LinkPreview.PersonMetadata.SchemaDotOrg;
-using Odin.Services.LinkPreview.Posts;
-using Odin.Services.LinkPreview.Profile;
+using Odin.Services.PublicPage.PersonMetadata.SchemaDotOrg;
+using Odin.Services.PublicPage.Posts;
+using Odin.Services.PublicPage.Profile;
 
-namespace Odin.Services.LinkPreview;
+namespace Odin.Services.PublicPage;
 
-public class LinkPreviewService(
+public class HomebasePublicPageService(
     ISystemLevel1Cache globalCache,
-    ITenantLevel1Cache<LinkPreviewService> tenantCache,
+    ITenantLevel1Cache<HomebasePublicPageService> tenantCache,
     HomebaseProfileContentService profileContentService,
     HomebaseChannelContentService channelContentService,
     IHttpContextAccessor httpContextAccessor,
     HomebaseSsrService ssrService,
-    ILogger<LinkPreviewService> logger)
+    ILogger<HomebasePublicPageService> logger)
 {
     private const string IndexFileKey = "link-preview-service-index-file";
     private const string GenericLinkPreviewCacheKey = "link-preview-service-index-file";
