+++ src/services/Odin.Services/PublicPage/PersonMetadata/SchemaDotOrg/AddressSchema.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Odin.Services.LinkPreview.PersonMetadata.SchemaDotOrg;
+namespace Odin.Services.PublicPage.PersonMetadata.SchemaDotOrg;
 
 public class AddressSchema
 {
