+++ src/services/Odin.Services/PublicPage/PersonMetadata/SchemaDotOrg/OrganizationSchema.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Odin.Services.LinkPreview.PersonMetadata.SchemaDotOrg;
+namespace Odin.Services.PublicPage.PersonMetadata.SchemaDotOrg;
 
 public class OrganizationSchema
 {
