+++ src/services/Odin.Services/PublicPage/Profile/ExperienceAttribute.cs
@@ -1,4 +1,4 @@
-namespace Odin.Services.LinkPreview.Profile;
+namespace Odin.Services.PublicPage.Profile;
 
 public class ExperienceAttribute
 {
